+++ llama_index/agent/openai_agent.py
@@ -1,16 +1,17 @@
 import asyncio
 import json
-import time
+import logging
 from abc import abstractmethod
+from enum import Enum
 from threading import Thread
-from typing import Any, Callable, List, Optional, Tuple, Type
+from typing import Any, Callable, Optional, Tuple, Type
 
 from llama_index.agent.types import BaseAgent
 from llama_index.callbacks.base import CallbackManager
 from llama_index.chat_engine.types import AgentChatResponse, StreamingAgentChatResponse
 from llama_index.indices.base_retriever import BaseRetriever
 from llama_index.indices.query.schema import QueryBundle
-from llama_index.llms.base import LLM, ChatMessage, MessageRole
+from llama_index.llms.base import LLM, ChatMessage, ChatResponse, MessageRole
 from llama_index.llms.openai import OpenAI
 from llama_index.llms.openai_utils import is_function_calling_model
 from llama_index.memory import BaseMemory, ChatMemoryBuffer
@@ -18,11 +19,14 @@ from llama_index.response.schema import RESPONSE_TYPE, Response
 from llama_index.schema import BaseNode, NodeWithScore
 from llama_index.tools import BaseTool, ToolOutput
 
+logger = logging.getLogger(__name__)
+logger.setLevel(logging.WARNING)
+
 DEFAULT_MAX_FUNCTION_CALLS = 5
 DEFAULT_MODEL_NAME = "gpt-3.5-turbo-0613"
 
 
-def get_function_by_name(tools: List[BaseTool], name: str) -> BaseTool:
+def get_function_by_name(tools: list[BaseTool], name: str) -> BaseTool:
     """Get function by name."""
     name_to_tool = {tool.metadata.name: tool for tool in tools}
     if name not in name_to_tool:
@@ -31,7 +35,7 @@ def get_function_by_name(tools: List[BaseTool], name: str) -> BaseTool:
 
 
 def call_function(
-    tools: List[BaseTool], function_call: dict, verbose: bool = False
+    tools: list[BaseTool], function_call: dict, verbose: bool = False
 ) -> Tuple[ChatMessage, ToolOutput]:
     """Call a function and return the output as a string."""
     name = function_call["name"]
@@ -57,267 +61,231 @@ def call_function(
     )
 
 
-class BaseOpenAIAgent(BaseAgent):
-    """Base OpenAI Agent."""
+class ChatMode(Enum):
+    default = 0
+    stream = 1
 
+
+class ChatSession:
+    def __init__(
+        self,
+        memory: BaseMemory,
+        prefix_messages: list[ChatMessage],
+        get_tools_callback: Callable[[str], list[BaseTool]],
+    ):
+        self.memory = memory
+        self.prefix_messages = prefix_messages
+        self.get_tools_callback = get_tools_callback
+        self.tools: list[BaseTool] = []
+        self.functions: list[dict[str, Any]] = []
+
+    @property
+    def chat_history(self) -> list[ChatMessage]:
+        return self.memory.get_all()
+
+    def reset(self) -> None:
+        self.memory.reset()
+
+    def prepare_message(self, message: str) -> Tuple[list[BaseTool], list[dict]]:
+        """Prepare tools and functions for the message."""
+        self.tools = self.get_tools_callback(message)
+        self.functions = [tool.metadata.to_openai_function() for tool in self.tools]
+        return self.tools, self.functions
+
+    def get_all_messages(self) -> list[ChatMessage]:
+        return self.prefix_messages + self.memory.get()
+
+    def get_latest_function_call(self) -> Optional[dict]:
+        return self.memory.get_all()[-1].additional_kwargs.get("function_call", None)
+
+
+class BaseOpenAIAgent(BaseAgent):
     def __init__(
         self,
         llm: OpenAI,
         memory: BaseMemory,
-        prefix_messages: List[ChatMessage],
-        verbose: bool = False,
-        max_function_calls: int = DEFAULT_MAX_FUNCTION_CALLS,
-        callback_manager: Optional[CallbackManager] = None,
-    ) -> None:
+        prefix_messages: list[ChatMessage],
+        verbose: bool,
+        max_function_calls: int,
+        callback_manager: Optional[CallbackManager],
+    ):
         self._llm = llm
-        self._memory = memory
-        self._prefix_messages = prefix_messages
         self._verbose = verbose
         self._max_function_calls = max_function_calls
         self.callback_manager = callback_manager or CallbackManager([])
+        self.session = ChatSession(memory, prefix_messages, self._get_tools)
+        self.sources: list[ToolOutput] = []
 
     @property
-    def chat_history(self) -> List[ChatMessage]:
-        return self._memory.get_all()
+    def chat_history(self) -> list[ChatMessage]:
+        return self.session.chat_history
+
+    @property
+    def all_messages(self) -> list[ChatMessage]:
+        return self.session.get_all_messages()
 
     def reset(self) -> None:
-        self._memory.reset()
+        self.session.reset()
 
     @abstractmethod
-    def _get_tools(self, message: str) -> List[BaseTool]:
+    def _get_tools(self, message: str) -> list[BaseTool]:
         """Get tools."""
+        pass
 
-    def _get_latest_function_call(
-        self, chat_history: List[ChatMessage]
-    ) -> Optional[dict]:
-        """Get latest function call from chat history."""
-        return chat_history[-1].additional_kwargs.get("function_call", None)
+    def _should_continue(
+        self, function_call: dict | None, n_function_calls: int
+    ) -> bool:
+        if n_function_calls > self._max_function_calls:
+            return False
+        if not function_call:
+            return False
+        return True
 
-    def _init_chat(self, message: str) -> Tuple[List[BaseTool], List[dict]]:
-        """Add user message to chat history and get tools and functions."""
-        self._memory.put(ChatMessage(content=message, role=MessageRole.USER))
-        tools = self._get_tools(message)
-        functions = [tool.metadata.to_openai_function() for tool in tools]
+    def init_chat(
+        self, message: str, chat_history: Optional[list[ChatMessage]] = None
+    ) -> tuple[list[BaseTool], list[dict]]:
+        if chat_history is not None:
+            self.session.memory.set(chat_history)
+        self.sources = []
+        self.session.memory.put(ChatMessage(content=message, role=MessageRole.USER))
+        tools, functions = self.session.prepare_message(message)
         return tools, functions
 
-    def chat(
-        self, message: str, chat_history: Optional[List[ChatMessage]] = None
-    ) -> AgentChatResponse:
-        if chat_history is not None:
-            self._memory.set(chat_history)
-
-        tools, functions = self._init_chat(message)
-        sources = []
-
-        # TODO: Support forced function call
-        all_messages = self._prefix_messages + self._memory.get()
-        chat_response = self._llm.chat(all_messages, functions=functions)
+    def _process_message(self, chat_response: ChatResponse) -> AgentChatResponse:
         ai_message = chat_response.message
-        self._memory.put(ai_message)
+        self.session.memory.put(ai_message)
+        return AgentChatResponse(response=str(ai_message.content), sources=self.sources)
 
-        n_function_calls = 0
-        function_call = self._get_latest_function_call(self._memory.get_all())
-        while function_call is not None:
-            if n_function_calls >= self._max_function_calls:
-                print(f"Exceeded max function calls: {self._max_function_calls}.")
-                break
-
-            function_message, tool_output = call_function(
-                tools, function_call, verbose=self._verbose
-            )
-            sources.append(tool_output)
-            self._memory.put(function_message)
-            n_function_calls += 1
-
-            # send function call & output back to get another response
-            all_messages = self._prefix_messages + self._memory.get()
-            chat_response = self._llm.chat(all_messages, functions=functions)
-            ai_message = chat_response.message
-            self._memory.put(ai_message)
-            function_call = self._get_latest_function_call(self._memory.get_all())
-
-        return AgentChatResponse(response=str(ai_message.content), sources=sources)
-
-    def stream_chat(
-        self, message: str, chat_history: Optional[List[ChatMessage]] = None
+    def _get_stream_ai_response(
+        self, functions: list[dict]
     ) -> StreamingAgentChatResponse:
-        if chat_history is not None:
-            self._memory.set(chat_history)
-        tools, functions = self._init_chat(message)
-        all_messages = self._prefix_messages + self._memory.get()
-        sources = []
-
-        # TODO: Support forced function call
         chat_stream_response = StreamingAgentChatResponse(
-            chat_stream=self._llm.stream_chat(all_messages, functions=functions)
+            chat_stream=self._llm.stream_chat(self.all_messages, functions=functions),
+            sources=self.sources,
         )
-
         # Get the response in a separate thread so we can yield the response
         thread = Thread(
             target=chat_stream_response.write_response_to_history,
-            args=(self._memory,),
+            args=(self.session.memory,),
         )
         thread.start()
+        # Wait for the event to be set
+        chat_stream_response._is_function_not_none_thread_event.wait()
+        # If it is executing an openAI function, wait for the thread to finish
+        if chat_stream_response._is_function:
+            thread.join()
+        # if it's false, return the answer (to stream)
+        return chat_stream_response
 
-        while chat_stream_response._is_function is None:
-            # Wait until we know if the response is a function call or not
-            time.sleep(0.05)
-            if chat_stream_response._is_function is False:
-                return chat_stream_response
+    async def _get_async_stream_ai_response(
+        self, functions: list[dict]
+    ) -> StreamingAgentChatResponse:
+        chat_stream_response = StreamingAgentChatResponse(
+            achat_stream=await self._llm.astream_chat(
+                self.all_messages, functions=functions
+            ),
+            sources=self.sources,
+        )
+        # create task to write chat response to history
+        asyncio.create_task(
+            chat_stream_response.awrite_response_to_history(self.session.memory)
+        )
+        # wait until openAI functions stop executing
+        await chat_stream_response._is_function_false_event.wait()
+        # return response stream
+        return chat_stream_response
 
-        thread.join()
+    def _call_function(self, tools: list[BaseTool], function_call: dict) -> None:
+        function_message, tool_output = call_function(
+            tools, function_call, verbose=self._verbose
+        )
+        self.sources.append(tool_output)
+        self.session.memory.put(function_message)
 
-        n_function_calls = 0
-        function_call = self._get_latest_function_call(self._memory.get_all())
-        while function_call is not None:
-            if n_function_calls >= self._max_function_calls:
-                print(f"Exceeded max function calls: {self._max_function_calls}.")
-                break
-
-            function_message, tool_output = call_function(
-                tools, function_call, verbose=self._verbose
+    def _get_agent_response(
+        self, mode: ChatMode, functions: list[dict]
+    ) -> AgentChatResponse | StreamingAgentChatResponse:
+        if mode == ChatMode.default:
+            chat_response: ChatResponse = self._llm.chat(
+                self.all_messages, functions=functions
             )
-            sources.append(tool_output)
-            self._memory.put(function_message)
+            return self._process_message(chat_response)
+        elif mode == ChatMode.stream:
+            return self._get_stream_ai_response(functions)
+        else:
+            raise NotImplementedError
+
+    async def _get_async_agent_response(
+        self, mode: ChatMode, functions: list[dict]
+    ) -> AgentChatResponse | StreamingAgentChatResponse:
+        if mode == ChatMode.default:
+            chat_response: ChatResponse = await self._llm.achat(
+                self.all_messages, functions=functions
+            )
+            return self._process_message(chat_response)
+        elif mode == ChatMode.stream:
+            return await self._get_async_stream_ai_response(functions)
+        else:
+            raise NotImplementedError
+
+    def chat(
+        self,
+        message: str,
+        chat_history: Optional[list[ChatMessage]] = None,
+        *,
+        mode: ChatMode = ChatMode.default,
+    ) -> AgentChatResponse | StreamingAgentChatResponse:
+        tools, functions = self.init_chat(message, chat_history)
+        n_function_calls = 0
+
+        # Loop until no more function calls or max_function_calls is reached
+        while True:
+            agent_chat_response = self._get_agent_response(mode, functions)
+            latest_function = self.session.get_latest_function_call()
+            if not self._should_continue(latest_function, n_function_calls):
+                logger.debug("Break: should continue False")
+                break
+            assert isinstance(latest_function, dict)
+            self._call_function(tools, latest_function)
             n_function_calls += 1
 
-            # send function call & output back to get another response
-            all_messages = self._prefix_messages + self._memory.get()
-            chat_stream_response = StreamingAgentChatResponse(
-                chat_stream=self._llm.stream_chat(all_messages, functions=functions),
-                sources=sources,
-            )
-
-            # Get the response in a separate thread so we can yield the response
-            thread = Thread(
-                target=chat_stream_response.write_response_to_history,
-                args=(self._memory,),
-            )
-            thread.start()
-            while chat_stream_response._is_function is None:
-                # Wait until we know if the response is a function call or not
-                time.sleep(0.05)
-                if chat_stream_response._is_function is False:
-                    return chat_stream_response
-
-            thread.join()
-            function_call = self._get_latest_function_call(self._memory.get_all())
-
-        return chat_stream_response
+        return agent_chat_response
 
     async def achat(
-        self, message: str, chat_history: Optional[List[ChatMessage]] = None
-    ) -> AgentChatResponse:
-        if chat_history is not None:
-            self._memory.set(chat_history)
-
-        tools, functions = self._init_chat(message)
-        sources = []
-
-        all_messages = self._prefix_messages + self._memory.get()
-
-        # TODO: Support forced function call
-        chat_response = await self._llm.achat(all_messages, functions=functions)
-        ai_message = chat_response.message
-        self._memory.put(ai_message)
-
+        self,
+        message: str,
+        chat_history: Optional[list[ChatMessage]] = None,
+        *,
+        mode: ChatMode = ChatMode.default,
+    ) -> AgentChatResponse | StreamingAgentChatResponse:
+        tools, functions = self.init_chat(message, chat_history)
         n_function_calls = 0
-        function_call = self._get_latest_function_call(self._memory.get_all())
-        while function_call is not None:
-            if n_function_calls >= self._max_function_calls:
-                print(f"Exceeded max function calls: {self._max_function_calls}.")
-                continue
 
-            function_message, tool_output = call_function(
-                tools, function_call, verbose=self._verbose
-            )
-            sources.append(tool_output)
-            self._memory.put(function_message)
+        # Loop until no more function calls or max_function_calls is reached
+        while True:
+            agent_chat_response = await self._get_async_agent_response(mode, functions)
+            latest_function = self.session.get_latest_function_call()
+            if not self._should_continue(latest_function, n_function_calls):
+                break
+            assert isinstance(latest_function, dict)
+            self._call_function(tools, latest_function)
             n_function_calls += 1
 
-            # send function call & output back to get another response
-            response = await self._llm.achat(
-                self._prefix_messages + self._memory.get(), functions=functions
-            )
-            ai_message = response.message
-            self._memory.put(ai_message)
-            function_call = self._get_latest_function_call(self._memory.get_all())
+        return agent_chat_response
 
-        return AgentChatResponse(response=str(ai_message.content), sources=sources)
+    def stream_chat(
+        self, message: str, chat_history: Optional[list[ChatMessage]] = None
+    ) -> StreamingAgentChatResponse:
+        chat_response = self.chat(message, chat_history, mode=ChatMode.stream)
+        assert isinstance(chat_response, StreamingAgentChatResponse)
+        return chat_response
 
     async def astream_chat(
-        self, message: str, chat_history: Optional[List[ChatMessage]] = None
+        self, message: str, chat_history: Optional[list[ChatMessage]] = None
     ) -> StreamingAgentChatResponse:
-        if chat_history is not None:
-            self._memory.set(chat_history)
-        tools, functions = self._init_chat(message)
-        all_messages = self._prefix_messages + self._memory.get()
-        sources = []
-
-        # TODO: Support forced function call
-        chat_stream_response = StreamingAgentChatResponse(
-            achat_stream=await self._llm.astream_chat(all_messages, functions=functions)
-        )
-
-        # Get the response in a separate thread so we can yield the response
-        thread = Thread(
-            target=lambda x: asyncio.run(
-                chat_stream_response.awrite_response_to_history(x)
-            ),
-            args=(self._memory,),
-        )
-        thread.start()
-
-        while chat_stream_response._is_function is None:
-            # Wait until we know if the response is a function call or not
-            time.sleep(0.05)
-            if chat_stream_response._is_function is False:
-                return chat_stream_response
-
-        thread.join()
-
-        n_function_calls = 0
-        function_call = self._get_latest_function_call(self._memory.get_all())
-        while function_call is not None:
-            if n_function_calls >= self._max_function_calls:
-                print(f"Exceeded max function calls: {self._max_function_calls}.")
-                break
-
-            function_message, tool_output = call_function(
-                tools, function_call, verbose=self._verbose
-            )
-            sources.append(tool_output)
-            self._memory.put(function_message)
-            n_function_calls += 1
-
-            # send function call & output back to get another response
-            all_messages = self._prefix_messages + self._memory.get()
-            chat_stream_response = StreamingAgentChatResponse(
-                achat_stream=await self._llm.astream_chat(
-                    all_messages, functions=functions
-                ),
-                sources=sources,
-            )
-
-            # Get the response in a separate thread so we can yield the response
-            thread = Thread(
-                target=lambda x: asyncio.run(
-                    chat_stream_response.awrite_response_to_history(x)
-                ),
-                args=(self._memory,),
-            )
-            thread.start()
-
-            while chat_stream_response._is_function is None:
-                # Wait until we know if the response is a function call or not
-                time.sleep(0.05)
-                if chat_stream_response._is_function is False:
-                    return chat_stream_response
-
-            thread.join()
-            function_call = self._get_latest_function_call(self._memory.get_all())
-
-        return chat_stream_response
+        chat_response = await self.achat(message, chat_history, mode=ChatMode.stream)
+        assert isinstance(chat_response, StreamingAgentChatResponse)
+        return chat_response
 
     # ===== Query Engine Interface =====
     def _query(self, query_bundle: QueryBundle) -> RESPONSE_TYPE:
@@ -338,10 +306,10 @@ class BaseOpenAIAgent(BaseAgent):
 class OpenAIAgent(BaseOpenAIAgent):
     def __init__(
         self,
-        tools: List[BaseTool],
+        tools: list[BaseTool],
         llm: OpenAI,
         memory: BaseMemory,
-        prefix_messages: List[ChatMessage],
+        prefix_messages: list[ChatMessage],
         verbose: bool = False,
         max_function_calls: int = DEFAULT_MAX_FUNCTION_CALLS,
         callback_manager: Optional[CallbackManager] = None,
@@ -359,18 +327,18 @@ class OpenAIAgent(BaseOpenAIAgent):
     @classmethod
     def from_tools(
         cls,
-        tools: Optional[List[BaseTool]] = None,
+        tools: Optional[list[BaseTool]] = None,
         llm: Optional[LLM] = None,
-        chat_history: Optional[List[ChatMessage]] = None,
+        chat_history: Optional[list[ChatMessage]] = None,
         memory: Optional[BaseMemory] = None,
         memory_cls: Type[BaseMemory] = ChatMemoryBuffer,
         verbose: bool = False,
         max_function_calls: int = DEFAULT_MAX_FUNCTION_CALLS,
         callback_manager: Optional[CallbackManager] = None,
         system_prompt: Optional[str] = None,
-        prefix_messages: Optional[List[ChatMessage]] = None,
+        prefix_messages: Optional[list[ChatMessage]] = None,
         **kwargs: Any,
-    ) -> "OpenAIAgent":
+    ) -> "ExperimentalOpenAIAgent":
         tools = tools or []
         chat_history = chat_history or []
         memory = memory or memory_cls.from_defaults(chat_history)
@@ -402,7 +370,7 @@ class OpenAIAgent(BaseOpenAIAgent):
             callback_manager=callback_manager,
         )
 
-    def _get_tools(self, message: str) -> List[BaseTool]:
+    def _get_tools(self, message: str) -> list[BaseTool]:
         """Get tools."""
         return self._tools
 
@@ -427,7 +395,7 @@ class RetrieverOpenAIAgent(BaseOpenAIAgent):
         node_to_tool_fn: Callable[[BaseNode], BaseTool],
         llm: OpenAI,
         memory: BaseMemory,
-        prefix_messages: List[ChatMessage],
+        prefix_messages: list[ChatMessage],
         verbose: bool = False,
         max_function_calls: int = DEFAULT_MAX_FUNCTION_CALLS,
         callback_manager: Optional[CallbackManager] = None,
@@ -449,14 +417,14 @@ class RetrieverOpenAIAgent(BaseOpenAIAgent):
         retriever: BaseRetriever,
         node_to_tool_fn: Callable[[BaseNode], BaseTool],
         llm: Optional[OpenAI] = None,
-        chat_history: Optional[List[ChatMessage]] = None,
+        chat_history: Optional[list[ChatMessage]] = None,
         memory: Optional[BaseMemory] = None,
         memory_cls: Type[BaseMemory] = ChatMemoryBuffer,
         verbose: bool = False,
         max_function_calls: int = DEFAULT_MAX_FUNCTION_CALLS,
         callback_manager: Optional[CallbackManager] = None,
         system_prompt: Optional[str] = None,
-        prefix_messages: Optional[List[ChatMessage]] = None,
+        prefix_messages: Optional[list[ChatMessage]] = None,
     ) -> "RetrieverOpenAIAgent":
         chat_history = chat_history or []
         memory = memory or memory_cls.from_defaults(chat_history)
@@ -490,12 +458,12 @@ class RetrieverOpenAIAgent(BaseOpenAIAgent):
             callback_manager=callback_manager,
         )
 
-    def _get_tools(self, message: str) -> List[BaseTool]:
-        retrieved_nodes_w_scores: List[NodeWithScore] = self._retriever.retrieve(
+    def _get_tools(self, message: str) -> list[BaseTool]:
+        retrieved_nodes_w_scores: list[NodeWithScore] = self._retriever.retrieve(
             message
         )
         retrieved_nodes = [node.node for node in retrieved_nodes_w_scores]
-        retrieved_tools: List[BaseTool] = [
+        retrieved_tools: list[BaseTool] = [
             self._node_to_tool_fn(n) for n in retrieved_nodes
         ]
         return retrieved_tools
